- display: removed standard library and ripped out required hardware functions. huge memory saving! - resolved issue with cossuped data on save by: 1) adding check sum 2) com_stop, com_star, - Nextion libarry bug - sendCommand function eats serial data - reported - display: added save message timeout 5 seconds - reset of EEPROM removed on startup and fixed button on display - display: increased display communication speed to 115200 - burner: amps values conversion fixed - display: added indication for ignition and pump UPS

+++ pits_burner.h
@@ -55,8 +55,8 @@ class PitsBurner
     byte getExhaustTemp();
     void setFeederTemp(byte);   //to overwrite sensors value
     byte getFeederTemp();
-    void setFeederAmps(float);   //to overwrite sensors value
-    float getFeederAmps();
+    void setFeederAmps(byte);   //to overwrite sensors value
+    byte getFeederAmps();
     byte getMinTemp();          //caclualted during heat start
     
     void setFlame(byte);        //to overwrite sensors value
@@ -138,7 +138,7 @@ class PitsBurner
     byte _intBattDVolts = 0;  
     PERCENT_RANGE _intFuelLevel = P0;
     byte _intFuelCm = 0;  
-    float _flFeedAmps[3] = {0,0,0};
+    byte _flFeedAmps[3] = {0,0,0};
     
     //object internal states
     unsigned int _intFeedTime = 0;            //total feed seconds counter, reset on power off
